+++ src/commands/welcome.ts
@@ -15,15 +15,19 @@ async function getPackages(context: ExtensionContext): Promise<PipPackage[]> {
 
 export default async function handle(context: ExtensionContext): Promise<void> {
     const silenced = context.globalState.get(SETTING_SILENCED);
+    console.log('silenced', silenced);
     if (silenced) { return; }
     const initialized = context.workspaceState.get(SETTING_INITIALIZED);
+    console.log('initialized', initialized);
     if (initialized) { return; }
 
     const pipPackages = await getPackages(context);
     const pipPackageNames = pipPackages.map(x => x.name);
+    console.log(pipPackageNames);
 
-    if (!pipPackageNames.includes('boto3')) { return; }
-    if (pipPackageNames.includes('boto3-stubs')) { return; }
+    if (!pipPackageNames.includes('boto3')) { console.log('boto3 not installed'); return; }
+    if (pipPackageNames.includes('boto3-stubs')) { console.log('boto3-stubs installed'); return; }
+
     const response = await window.showInformationMessage(
         `This project uses boto3 with no ${NAME}.`,
         `Install ${NAME}`,
@@ -34,7 +38,7 @@ export default async function handle(context: ExtensionContext): Promise<void> {
     if (response === 'Do not show this again') {
         context.globalState.update(SETTING_SILENCED, 'true');
     }
-    if (response === 'Install it now') {
+    if (response === `Install ${NAME}`) {
         await quickstartHandler(context);
     }
     if (response === 'Maybe later') {